OS Windows support

+++ engine/gui/Button.cpp
@@ -46,14 +46,14 @@ void Button::press()
 void Button::init() {
     _button.setTexture(*ResourceManager::loadTexture(_texture));
     _button.setTextureRect(sf::IntRect(_usualState.tx, _usualState.ty, _w, _h));
-    _button.scale(_sx, _sy);
-    _button.setPosition(_x - _w * _sx / 2, _y - _h * _sy / 2);
+    _button.scale((float)_sx, (float)_sy);
+    _button.setPosition((float)(_x - _w * _sx / 2), (float)(_y - _h * _sy / 2));
 
     _text.setFont(*ResourceManager::loadFont(_font));
     _text.setString(_textString);
-    _text.setCharacterSize(_h * _sy / 2);
+    _text.setCharacterSize((unsigned int)(_h * _sy / 2));
     _text.setFillColor(_textColor);
-    _text.setPosition(_x - _text.getLocalBounds().width / 2, _y - _h * _sy / 2 + _text.getLocalBounds().height / 4);
+    _text.setPosition((float)(_x - _text.getLocalBounds().width / 2), (float)(_y - _h * _sy / 2 + _text.getLocalBounds().height / 4));
 
     _clickSound.setBuffer(*ResourceManager::loadSoundBuffer(_clickSoundName));
     _clickSound.setVolume(15);
